Trial run of using memory array of uint64_t
May help with node offset issues in hw_emu, untested

+++ src/krnl.hpp
@@ -62,9 +62,9 @@ void krnl(
 	int RDMA_TYPE,
 	int exec,
 	//! [inout] Pointer to on-chip high-bandwidth memory
-	uint8_t *hbm,
-	uint8_t *req_buffer,
-	uint8_t *resp_buffer,
+	uint64_t *hbm,
+	uint64_t *req_buffer,
+	uint64_t *resp_buffer,
 	//! [inout] Address within HBM that holds the root of the tree
 	//!
 	//! Can be modified by operation kernels
