+++ src/tests/insert/until-it-breaks.hpp
@@ -21,9 +21,9 @@ bool until_it_breaks(
 	int myBoardNum,
 	int RDMA_TYPE,
 	int exec,
-	uint8_t *hbm,
-	uint8_t *req_buffer,
-	uint8_t *resp_buffer
+	uint64_t *hbm,
+	uint64_t *req_buffer,
+	uint64_t *resp_buffer
 );
